Removes Firebase messaging and storage
Removes Firebase messaging and storage dependencies and related code.

This change simplifies the application and reduces the number of external dependencies.  The `di.dart` file was also deleted and replaced with `dependency_injection.dart`, likely due to a significant refactoring or change in dependency management approach.  Removed the manual JSON serialization logic from the `UserEntity`.