Refactors BaseScreen and adds Preferences screen
Refactors BaseScreen to streamline its structure and improve code readability.  Moves protected method declarations closer to their usage. Adds a GestureDetector to the body to dismiss the keyboard on tap.

Adds a new Preferences screen with basic navigation. Includes new SVG assets for logout and visibility icons.  Updates the Account screen UI to accommodate post titles and routing to the Preferences screen.  Modifies RootController to only fetch user data if the user is signed in.  Updates AppPages to include the new Preferences and Splash screens, making Splash the initial route.